FDG-6272 Refactored key takeaways
