+++ src/layouts/explainer/sections/national-debt/national-debt.jsx
@@ -49,14 +49,10 @@ import { pxToNumber } from "../../../../helpers/styles-helper/styles-helper";
 import curvedArrow from "../../../../images/curved-arrow.svg";
 import alexanderHamilton from "../../../../images/alexander-hamilton.png";
 import benFranklin from "../../../../images/ben-franklin.png";
+import { KeyTakeawaysSection } from "./key-takeaways/national-debt-key-takeaways";
 
 import {
-  // Key Takeaways
-  keyTakeawaysContent,
   icon,
-  offsetIcon,
-  iconBackground,
-  noMarginBottom,
   // NationalDebtExplained
   nationalDebtExplainedTextContent,
   nationalDebtExplainedTable,
@@ -166,68 +162,6 @@ export const visualizingTheDebtTableContent = {
 
 export const chartPatternBackground = "#4A0072";
 const alternateBarColor = "#b699c6";
-
-const KeyTakeawaysSection = ({ glossary }) => {
-  const nonMarketableSecurities = (
-    <GlossaryTerm
-      term="Non-Marketable Securities"
-      page="Debt explainer"
-      glossary={glossary}
-    >
-      non-marketable
-    </GlossaryTerm>
-  );
-  const marketableSecurities = (
-    <GlossaryTerm
-      term="Marketable Securities"
-      page="Debt explainer"
-      glossary={glossary}
-    >
-      marketable
-    </GlossaryTerm>
-  );
-
-  return (
-    <>
-      <div className={keyTakeawaysContent}>
-        <div className={iconBackground}>
-          <FontAwesomeIcon icon={faMoneyCheckDollar} className={icon} />
-          <FontAwesomeIcon icon={faMoneyCheckDollar} className={offsetIcon} />
-        </div>
-        <p>
-          The national debt is composed of distinct types of debt, similar to an
-          individual whose debt may consist of a mortgage, car loan, and credit
-          cards. The different types of debt include {nonMarketableSecurities}{" "}
-          or {marketableSecurities} securities and whether it is debt held by
-          the public or debt held by the government itself (known as
-          intragovernmental).
-        </p>
-      </div>
-      <div className={keyTakeawaysContent}>
-        <div className={iconBackground}>
-          <FontAwesomeIcon icon={faChartLine} className={icon} />
-          <FontAwesomeIcon icon={faChartLine} className={offsetIcon} />
-        </div>
-        <p>
-          The U.S. has carried debt since its inception. Debts incurred during
-          the American Revolutionary War amounted to $75 million, primarily
-          borrowed from domestic investors and the French Government for war
-          materials.
-        </p>
-      </div>
-      <div className={`${keyTakeawaysContent} ${noMarginBottom}`}>
-        <div className={iconBackground}>
-          <FontAwesomeIcon icon={faPeopleCarry} className={icon} />
-          <FontAwesomeIcon icon={faPeopleCarry} className={offsetIcon} />
-        </div>
-        <p>
-          The national debt enables the federal government to pay for important
-          programs and services for the American public.
-        </p>
-      </div>
-    </>
-  );
-};
 
 export const NationalDebtExplainedSection = ({ glossary, cpiDataByYear }) => {
   const glossaryTerms = {
@@ -1795,7 +1729,7 @@ export const DebtBreakdownSection = withWindowSize(
             </>
           )}
         </div>
-        <div className={postGraphContent}>
+        <div className={postGraphContent} id={'interest-rate-and-total-debt-chart'}>
           <h3>Maintaining the National Debt</h3>
           <p>
             The federal government is charged interest for the use of lenders’
@@ -1817,7 +1751,7 @@ export const DebtBreakdownSection = withWindowSize(
             in interest rates and inflation are now resulting in an increase in
             interest expense.
           </p>
-          <div className={visWithCallout} id={'interest-rate-and-total-debt-chart'}>
+          <div className={visWithCallout}>
             {multichartDataLoaded && (
               <div className={multichartWrapper}>
                 <div
